% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_comp_n_stand.R
\name{comp.n.q.pad}
\alias{comp.n.q.pad}
\title{Function to create the plot of theta stan comp n}
\usage{
comp.n.q.pad(values, n.values, v.values)
}
\description{
Function to create the plot of theta stan comp n
}
