% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_asymp.R
\name{prob.asym}
\alias{prob.asym}
\title{Analyze the value of n using the theoretical quantities}
\usage{
prob.asym(n, mu, theta, v, u = -v)
}
\description{
Analyze the value of n using the theoretical quantities
}
