% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qexp_generate_functions.R
\name{qexp.samples}
\alias{qexp.samples}
\title{qexp.samples}
\usage{
qexp.samples(N, n, theta, mu)
}
\arguments{
\item{N}{Numeric. Number of samples}

\item{n}{Numeric. Sample size of each sample}

\item{theta}{Numeric. True value of theta}

\item{mu}{Numeric. True value of mu}
}
\value{

}
\description{
Generate samples
}
